requirements.txt
networkx==3.2.1
python-dotenv==1.0.0
openai==1.12.0 